+++ src/main/java/brig/concord/psi/impl/delegate/ConcordYamlDelegateFactory.java
@@ -1,7 +1,6 @@
 package brig.concord.psi.impl.delegate;
 
 import brig.concord.meta.ConcordMetaTypeProvider;
-import brig.concord.meta.ConcordMetaTypeProxy;
 import brig.concord.meta.model.CallMetaType;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.lang.ASTNode;
@@ -52,18 +51,11 @@ public class ConcordYamlDelegateFactory {
 
     private static PsiNamedElement createPlainTextDelegate(YAMLPlainTextImpl yamlPlainText) {
         ConcordMetaTypeProvider instance = ConcordMetaTypeProvider.getInstance(yamlPlainText.getProject());
-        YamlMetaType metaType = unwindMetaType(instance.getResolvedMetaType(yamlPlainText));
+        YamlMetaType metaType = instance.getResolvedMetaType(yamlPlainText);
         if (metaType instanceof CallMetaType) {
             return new YamlFlowCallDelegate(yamlPlainText);
         }
         return null;
-    }
-
-    private static YamlMetaType unwindMetaType(YamlMetaType metaType) {
-        if (metaType instanceof ConcordMetaTypeProxy.YamlMetaTypeProxy proxy) {
-            return proxy.original();
-        }
-        return metaType;
     }
 
     private static class NotADelegate extends ASTWrapperPsiElement implements PsiNamedElement {