+++ src/main/java/brig/concord/ConcordBundle.java
@@ -6,8 +6,6 @@ import org.jetbrains.annotations.NonNls;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.PropertyKey;
 
-import java.util.function.Supplier;
-
 public final class ConcordBundle extends DynamicBundle {
     @NonNls
     public static final String BUNDLE = "messages.ConcordBundle";
@@ -20,10 +18,5 @@ public final class ConcordBundle extends DynamicBundle {
     @NotNull
     public static @Nls String message(@NotNull @PropertyKey(resourceBundle = BUNDLE) String key, Object @NotNull ... params) {
         return INSTANCE.getMessage(key, params);
-    }
-
-    @NotNull
-    public static Supplier<@Nls String> messagePointer(@NotNull @PropertyKey(resourceBundle = BUNDLE) String key, Object @NotNull ... params) {
-        return INSTANCE.getLazyMessage(key, params);
     }
 }