+++ src/main/java/brig/concord/documentation/Documented.java
@@ -1,0 +1,5 @@
+package brig.concord.documentation;
+
+public interface Documented {
+    String getDescription();
+}