+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Support for `return` step
+
 ## [0.0.6 - 2022-11-16]
 
 ### Changed